ai-agent/requirements.txt
# Casper SDK
pycspr>=0.14.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Machine Learning
scikit-learn>=1.3.0

# HTTP Client
httpx>=0.25.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Configuration
pyyaml>=6.0.1

# Async utilities
asyncio-throttle>=1.0.2

# Logging
structlog>=23.2.0
